Calculate relative abundance

Calculates the expected abundance relative to zero exploitation, years.vec years after projections start

abund_rel(traj, zero.traj, K = NA, years.vec = c(10, 20, 50), fulldist = TRUE)

Arguments

traj
a matrix of trajectories, with rows=nsims and cols=nyears. Can be produced from projections()

zero.traj
a matrix of trajectories, same as traj but unfished. CHECK to make sure both start at the same simulation, in the app code.

K
if calculating abundance relative to K, put in K.

years.vec
a vector of years to check abundances at. E.g., how many years after the start of projections do you want to know abundance?

fulldist
logical saying whether to return the full distribution of relative abundances or the median of all the relative abundances.

Value

a vector of abundance relative to K or zero-exploitation where nrows=length of years.vec (10 years after, 20 years, etc.) and ncol=number of simulations.

Examples

parms <- list(
  S0 = 0.944, S1plus = 0.99,
  K1plus = 9000,
  AgeMat = 18, nages = 20,
  z = 2.39, lambdaMax = 1.02
)
initdepl <- 0.5
traj <- projections(
  NOut = 50,
  ConstantBycatch = list(
    Catch = 50,
    CV = 0.3
  ),
  InitDepl = initdepl,
  lh.params = parms,
  nyears = 100
)$trajectories
traj0 <- projections(
  NOut = 50,
  ConstantBycatch = list(
    Catch = 0,
    CV = 0
  ),
  InitDepl = initdepl,
  lh.params = parms,
  nyears = 100
)$trajectories
abund_rel(traj = traj, zero.traj = traj0, K = parms$K1plus)

#> $YrPost10
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5232541 0.5342939 0.5188403 0.5323365 0.5311041 0.5321085 0.5295617 0.5326711 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5178932 0.5229265 0.5242369 0.5305502 0.5236245 0.5309808 0.5247784 0.5209671 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5195121 0.5209885 0.5235504 0.5266269 0.5255319 0.5272642 0.5208391 0.5181160 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5335013 0.5283214 0.5235851 0.5351307 0.5254234 0.5257992 0.5223248 0.5255028 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5258969 0.5213085 0.5295491 0.5283035 0.5224774 0.5281776 0.5237174 0.5244107 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5243985 0.5257539 0.5279985 0.5167761 0.5234849 0.5182496 0.5239998 0.5259923 
#>      traj      traj 
#> 0.5322987 0.5299032 
#> 
#> $YrPost20
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5490998 0.5675638 0.5494357 0.5568112 0.5628141 0.5655177 0.5660071 0.5554411 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5517957 0.5472208 0.5440196 0.5511449 0.5613857 0.5617840 0.5527922 0.5487220 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5447372 0.5444469 0.5571158 0.5482326 0.5572957 0.5601999 0.5474146 0.5476654 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5714847 0.5502517 0.5564479 0.5667897 0.5579626 0.5612884 0.5463950 0.5613815 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5424682 0.5530460 0.5585828 0.5581131 0.5539004 0.5579258 0.5582970 0.5559422 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.5509786 0.5617662 0.5606308 0.5555079 0.5475267 0.5459461 0.5547733 0.5608205 
#>      traj      traj 
#> 0.5584892 0.5554121 
#> 
#> $YrPost50
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.6492392 0.6709321 0.6518742 0.6506366 0.6549589 0.6514703 0.6665326 0.6326392 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.6503508 0.6380351 0.6300217 0.6509257 0.6624535 0.6632731 0.6545834 0.6358923 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.6259238 0.6482648 0.6551920 0.6488893 0.6513472 0.6689994 0.6480924 0.6515420 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.6698655 0.6477677 0.6533427 0.6609406 0.6475267 0.6606529 0.6534510 0.6739644 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.6391034 0.6450283 0.6575797 0.6496220 0.6556595 0.6704395 0.6401315 0.6321488 
#>      traj      traj      traj      traj      traj      traj      traj      traj 
#> 0.6516518 0.6604912 0.6643887 0.6620023 0.6376882 0.6270492 0.6376130 0.6606742 
#>      traj      traj 
#> 0.6618461 0.6571658 
#>