Calculate normalized sustainable yield

This function calculates the normalized sustainable yield, which is used to find MNPL (the population size at which productivity is maximized).

ce(
  S0 = NA,
  S1plus = NA,
  AgeMat = NA,
  nages = NA,
  z = NA,
  E = NA,
  A = NA,
  P0 = NA,
  N0 = NA
)

Arguments

S0
Calf/pup survival, a numeric value between 0 and 1

S1plus
1+ survival rate for animals age 1 year and older, a numeric value between 0 and 1

AgeMat
Age at maturity (= age at first parturition - 1). Must be less than nages

nages
"maximum" age, treated as the plus group age. The plus group age can be set equal to the age at maturity +2 years without losing accuracy.

z
degree of compensation

E
bycatch mortality rate (applies to 1+ numbers)

A
the Pella-Tomlinson resilience parameter ((fmax - f0)/f0)

P0
unfished number-per-recruit - 1+ adults

N0
unfished numbers-per-recruit - mature adults

Value

a single value of normalized yield for exploitation rate E

Examples

# Set parameters
S0.w = 0.5; S1plus.w = 0.944; nages.w = 25; AgeMat.w = 18 
# Get number of individuals per recruit in terms of mature individuals (N0.w)
NPROut <- npr(S0 = S0.w, S1plus = S1plus.w, nages = nages.w, AgeMat = AgeMat.w, E = 0)

N0 <- NPROut$npr # mature numbers per recruit
# Get number of individuals per recruit in terms of individuals aged 1+ (P0.w)
P0 <- NPROut$P1r # 1+ nums per recruit

ce(S0 = S0.w, S1plus = S1plus.w, 
nages = nages.w, 
AgeMat = AgeMat.w, 
E=0.01, z=2.39,A=2, N0 = N0, P0 = P0)

#> [1] 0.08132041