Generate one marine mammal population trajectory

This function generates one trajectory for a marine mammal population, starting at a user-specified depletion level InitDepl.

dynamics(
  S0,
  S1plus,
  K1plus,
  AgeMat,
  InitDepl,
  ConstantCatch = NA,
  ConstantF = NA,
  z,
  nyears,
  nages,
  lambdaMax
)

Arguments

S0
Calf/pup survival, a numeric value between 0 and 1

S1plus
Survival for animals age 1 year and older, a numeric value between 0 and 1

K1plus
The pre-exploitation population size of individuals aged 1 and older. If this value is unavailable, it can be approximated by using the initial depletion and the estimate of current abundance

AgeMat
Age at maturity in years (assumed to be age at first parturition - 1).

InitDepl
Starting depletion level

ConstantCatch
Total bycatch each year, expressed as a vector of length nyears

ConstantF
vector (length = nyears) rate of bycatch each year

z
The degree of compensation. The default value is z = 2.39.

nyears
Number of years to project

nages
"Maximum" age, treated as the plus group age. The plus group age can be set equal to the age at maturity +2 years without losing accuracy. Must be greater than AgeMat.

lambdaMax
Maximum steady rate of increase (population growth rate)

Value

A list containing a matrix N of numbers at age (dimensions nyears (rows) x nages (columns)) and one vector TotalPop (a vector of length nyears), containing the number of age 1+ individuals in the population.

Details

The population model is a single-sex age-structured model in which the number of calves or pups born each year is density dependent, with the extent of density dependence a function of the number of mature adults \(\tildeN\), the fecundity (pregnancy rate) at pre-exploitation equilibrium \(f_0\), the maximum theoretical fecundity rate fmax, the degree of compensation \(z\), and the abundance of individuals aged 1+ \(N_{y+1}^{1+}\) relative to carrying capacity \(K^{1+}\). This function can be used alone but is intended to be used with Projections() to generate multiple simulations. NOTE: Either ConstantCatch or ConstantF can be specified, but not both.

Examples

# Generate a time series of abundance for a bowhead whale
dynamics(S0 = 0.944, S1plus = 0.99, K1plus = 9000, AgeMat = 17,
 InitDepl = 0.6, ConstantCatch = NA, ConstantF = rep(0.01, times = 100), 
 z = 2.39, nyears = 100, nages = 25, lambdaMax = 1.04)

#> $TotalPop
#>   [1] 5400.000 5510.076 5618.662 5725.663 5830.986 5934.542 6036.247 6136.018
#>   [9] 6233.781 6329.466 6423.009 6514.353 6603.446 6690.247 6774.720 6856.836
#>  [17] 6936.578 7013.931 7088.894 7161.290 7231.009 7297.966 7362.102 7423.382
#>  [25] 7481.794 7537.345 7590.062 7639.991 7687.189 7731.728 7773.691 7813.168
#>  [33] 7850.257 7885.062 7917.687 7948.241 7976.833 8003.565 8028.538 8051.847
#>  [41] 8073.585 8093.841 8112.700 8130.246 8146.556 8161.708 8175.772 8188.820
#>  [49] 8200.917 8212.126 8222.507 8232.117 8241.008 8249.232 8256.837 8263.867
#>  [57] 8270.364 8276.367 8281.913 8287.036 8291.767 8296.136 8300.169 8303.892
#>  [65] 8307.328 8310.499 8313.425 8316.125 8318.615 8320.912 8323.030 8324.983
#>  [73] 8326.784 8328.445 8329.976 8331.387 8332.687 8333.886 8334.991 8336.009
#>  [81] 8336.948 8337.813 8338.610 8339.345 8340.022 8340.646 8341.221 8341.750
#>  [89] 8342.239 8342.688 8343.103 8343.485 8343.837 8344.161 8344.460 8344.735
#>  [97] 8344.988 8345.222 8345.437 8345.635
#> 
#> $N
#>            [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>   [1,] 230.4410 217.5363 208.7730 200.3627 192.2912 184.5448 177.1105 169.9757
#>   [2,] 231.1828 217.5363 213.2074 204.6184 196.3755 188.4646 180.8724 173.5860
#>   [3,] 231.7924 218.2365 213.2074 208.9645 200.5465 192.4676 184.7141 177.2730
#>   [4,] 232.2709 218.8121 213.8936 208.9645 204.8062 196.5556 188.6375 181.0383
#>   [5,] 232.6195 219.2637 214.4577 209.6371 204.8062 200.7305 192.6442 184.8836
#>   [6,] 232.8407 219.5928 214.9003 210.1900 205.4654 200.7305 196.7360 188.8106
#>   [7,] 232.9374 219.8017 215.2230 210.6238 206.0072 201.3766 196.7360 192.8209
#>   [8,] 232.9132 219.8929 215.4276 210.9400 206.4324 201.9077 197.3692 192.8209
#>   [9,] 232.7725 219.8701 215.5170 211.1406 206.7423 202.3244 197.8897 193.4415
#>  [10,] 232.5204 219.7373 215.4947 211.2282 206.9389 202.6281 198.2981 193.9517
#>  [11,] 232.1624 219.4992 215.3645 211.2063 207.0248 202.8208 198.5958 194.3520
#>  [12,] 231.7047 219.1613 215.1312 211.0787 207.0033 202.9050 198.7847 194.6438
#>  [13,] 231.1540 218.7292 214.8000 210.8501 206.8783 202.8839 198.8672 194.8289
#>  [14,] 230.5175 218.2094 214.3765 210.5254 206.6542 202.7614 198.8466 194.9097
#>  [15,] 229.8027 217.6085 213.8670 210.1104 206.3360 202.5417 198.7264 194.8895
#>  [16,] 229.0172 216.9337 213.2781 209.6111 205.9292 202.2299 198.5112 194.7718
#>  [17,] 228.1692 216.1922 212.6167 209.0339 205.4398 201.8312 198.2055 194.5608
#>  [18,] 227.2667 215.3917 211.8900 208.3857 204.8741 201.3516 197.8148 194.2612
#>  [19,] 226.1282 214.5397 211.1054 207.6734 204.2388 200.7971 197.3447 193.8783
#>  [20,] 224.8184 213.4651 210.2704 206.9044 203.5407 200.1744 196.8013 193.4175
#>  [21,] 223.3625 212.2286 209.2171 206.0860 202.7870 199.4903 196.1910 192.8849
#>  [22,] 221.7857 210.8542 208.0052 205.0537 201.9849 198.7516 195.5204 192.2868
#>  [23,] 220.1122 209.3657 206.6582 203.8659 200.9731 197.9654 194.7964 191.6295
#>  [24,] 218.3652 207.7859 205.1994 202.5457 199.8090 196.9737 194.0259 190.9199
#>  [25,] 216.5663 206.1367 203.6510 201.1159 198.5151 195.8328 193.0540 190.1648
#>  [26,] 214.7359 204.4386 202.0346 199.5984 197.1137 194.5646 191.9357 189.2122
#>  [27,] 212.8925 202.7107 200.3703 198.0141 195.6263 193.1911 190.6928 188.1162
#>  [28,] 211.0527 200.9706 198.6768 196.3829 194.0736 191.7334 189.3466 186.8980
#>  [29,] 209.2312 199.2338 196.9712 194.7231 192.4749 190.2116 187.9179 185.5786
#>  [30,] 207.4408 197.5143 195.2690 193.0515 190.8481 188.6446 186.4263 184.1783
#>  [31,] 205.6925 195.8242 193.5837 191.3832 189.2098 187.0503 184.8906 182.7165
#>  [32,] 203.9953 194.1737 191.9273 189.7314 187.5746 185.4445 183.3280 181.2113
#>  [33,] 202.3565 192.5715 190.3097 188.1079 185.9558 183.8419 181.7542 179.6797
#>  [34,] 200.7819 191.0245 188.7394 186.5225 184.3646 182.2552 180.1834 178.1373
#>  [35,] 199.2757 189.5381 187.2231 184.9834 182.8107 180.6957 178.6284 176.5978
#>  [36,] 197.8408 188.1162 185.7663 183.4974 181.3023 179.1728 177.0999 175.0737
#>  [37,] 196.4740 186.7617 184.3727 182.0695 179.8458 177.6944 175.6072 173.5756
#>  [38,] 195.1735 185.4714 183.0451 180.7037 178.4463 176.2669 174.1582 172.1126
#>  [39,] 193.9375 184.2438 181.7806 179.4025 177.1077 174.8953 172.7592 170.6925
#>  [40,] 192.7645 183.0770 180.5773 178.1631 175.8324 173.5833 171.4148 169.3212
#>  [41,] 191.6526 181.9697 179.4338 176.9838 174.6177 172.3334 170.1289 168.0037
#>  [42,] 190.6004 180.9201 178.3485 175.8631 173.4618 171.1428 168.9039 166.7434
#>  [43,] 189.6060 179.9268 177.3198 174.7993 172.3634 170.0099 167.7370 165.5427
#>  [44,] 188.6677 178.9881 176.3462 173.7911 171.3208 168.9333 166.6267 164.3991
#>  [45,] 187.7838 178.1023 175.4262 172.8369 170.3327 167.9115 165.5716 163.3109
#>  [46,] 186.9523 177.2679 174.5581 171.9352 169.3975 166.9431 164.5701 162.2767
#>  [47,] 186.1714 176.4830 173.7402 171.0844 168.5137 166.0265 163.6209 161.2951
#>  [48,] 185.4391 175.7458 172.9710 170.2828 167.6798 165.1603 162.7225 160.3648
#>  [49,] 184.7535 175.0545 172.2485 169.5288 166.8942 164.3430 161.8736 159.4844
#>  [50,] 184.1124 174.4073 171.5709 168.8207 166.1552 163.5730 161.0725 158.6523
#>  [51,] 183.5140 173.8022 170.9366 168.1567 165.4612 162.8487 160.3179 157.8672
#>  [52,] 182.9561 173.2372 170.3435 167.5349 164.8104 162.1685 159.6080 157.1276
#>  [53,] 182.4367 172.7106 169.7898 166.9537 164.2010 161.5306 158.9414 156.4318
#>  [54,] 181.9538 172.2202 169.2736 166.4110 163.6313 160.9334 158.3162 155.7784
#>  [55,] 181.5053 171.7644 168.7931 165.9051 163.0994 160.3750 157.7308 155.1657
#>  [56,] 181.0890 171.3410 168.3463 165.4341 162.6036 159.8537 157.1836 154.5920
#>  [57,] 180.7030 170.9480 167.9313 164.9962 162.1419 159.3678 156.6726 154.0556
#>  [58,] 180.3452 170.5836 167.5462 164.5894 161.7128 158.9153 156.1963 153.5549
#>  [59,] 180.0137 170.2458 167.1890 164.2120 161.3141 158.4947 155.7529 153.0880
#>  [60,] 179.7066 169.9329 166.8579 163.8619 160.9442 158.1040 155.3406 152.6534
#>  [61,] 179.4224 169.6431 166.5512 163.5375 160.6011 157.7414 154.9577 152.2493
#>  [62,] 179.1592 169.3747 166.2672 163.2369 160.2831 157.4051 154.6023 151.8740
#>  [63,] 178.9157 169.1263 166.0042 162.9584 159.9884 157.0934 154.2728 151.5257
#>  [64,] 178.6904 168.8964 165.7607 162.7007 159.7156 156.8047 153.9673 151.2027
#>  [65,] 178.4819 168.6837 165.5354 162.4621 159.4629 156.5372 153.6843 150.9033
#>  [66,] 178.2892 168.4870 165.3269 162.2412 159.2291 156.2896 153.4221 150.6259
#>  [67,] 178.1109 168.3050 165.1341 162.0369 159.0126 156.0604 153.1795 150.3690
#>  [68,] 177.9460 168.1367 164.9557 161.8479 158.8124 155.8483 152.9548 150.1312
#>  [69,] 177.7937 167.9811 164.7907 161.6731 158.6271 155.6520 152.7469 149.9110
#>  [70,] 177.6529 167.8372 164.6382 161.5114 158.4558 155.4704 152.5545 149.7072
#>  [71,] 177.5228 167.7043 164.4973 161.3619 158.2973 155.3025 152.3766 149.5187
#>  [72,] 177.4027 167.5816 164.3670 161.2238 158.1508 155.1472 152.2120 149.3443
#>  [73,] 177.2918 167.4681 164.2467 161.0961 158.0154 155.0036 152.0598 149.1830
#>  [74,] 177.1894 167.3634 164.1355 160.9782 157.8903 154.8709 151.9191 149.0338
#>  [75,] 177.0949 167.2668 164.0329 160.8692 157.7747 154.7483 151.7890 148.8959
#>  [76,] 177.0077 167.1776 163.9382 160.7686 157.6679 154.6350 151.6688 148.7684
#>  [77,] 176.9273 167.0953 163.8507 160.6758 157.5694 154.5303 151.5578 148.6506
#>  [78,] 176.8531 167.0193 163.7701 160.5901 157.4783 154.4337 151.4552 148.5418
#>  [79,] 176.7847 166.9493 163.6957 160.5110 157.3944 154.3445 151.3605 148.4412
#>  [80,] 176.7216 166.8847 163.6270 160.4381 157.3169 154.2622 151.2731 148.3484
#>  [81,] 176.6634 166.8252 163.5637 160.3708 157.2454 154.1863 151.1924 148.2627
#>  [82,] 176.6098 166.7703 163.5054 160.3088 157.1795 154.1162 151.1180 148.1837
#>  [83,] 176.5604 166.7197 163.4516 160.2516 157.1187 154.0516 151.0493 148.1107
#>  [84,] 176.5148 166.6730 163.4019 160.1989 157.0626 153.9920 150.9860 148.0434
#>  [85,] 176.4728 166.6300 163.3562 160.1502 157.0109 153.9371 150.9276 147.9813
#>  [86,] 176.4341 166.5903 163.3140 160.1054 156.9633 153.8864 150.8737 147.9241
#>  [87,] 176.3984 166.5538 163.2752 160.0641 156.9193 153.8397 150.8241 147.8713
#>  [88,] 176.3655 166.5201 163.2393 160.0260 156.8788 153.7966 150.7783 147.8227
#>  [89,] 176.3351 166.4890 163.2063 159.9909 156.8415 153.7569 150.7361 147.7778
#>  [90,] 176.3072 166.4604 163.1759 159.9585 156.8071 153.7203 150.6972 147.7364
#>  [91,] 176.2814 166.4340 163.1478 159.9287 156.7753 153.6866 150.6613 147.6983
#>  [92,] 176.2576 166.4096 163.1219 159.9012 156.7461 153.6555 150.6282 147.6631
#>  [93,] 176.2357 166.3872 163.0981 159.8758 156.7191 153.6268 150.5978 147.6307
#>  [94,] 176.2156 166.3665 163.0761 159.8524 156.6943 153.6004 150.5697 147.6009
#>  [95,] 176.1970 166.3475 163.0558 159.8309 156.6714 153.5761 150.5438 147.5733
#>  [96,] 176.1798 166.3299 163.0372 159.8110 156.6503 153.5536 150.5199 147.5479
#>  [97,] 176.1640 166.3138 163.0200 159.7927 156.6308 153.5329 150.4979 147.5245
#>  [98,] 176.1495 166.2988 163.0041 159.7759 156.6129 153.5138 150.4776 147.5030
#>  [99,] 176.1361 166.2851 162.9895 159.7603 156.5963 153.4963 150.4589 147.4831
#> [100,] 176.1237 166.2725 162.9760 159.7460 156.5811 153.4801 150.4417 147.4648
#>            [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#>   [1,] 163.1283 156.5568 150.2500 144.1972 138.3883 132.8134 127.4631 122.3283
#>   [2,] 166.5932 159.8821 153.4413 147.2600 141.3277 135.6344 130.1704 124.9266
#>   [3,] 170.1317 163.2780 156.7004 150.3878 144.3295 138.5153 132.9353 127.5800
#>   [4,] 173.7453 166.7460 160.0288 153.5821 147.3951 141.4574 135.7588 130.2898
#>   [5,] 177.4357 170.2877 163.4278 156.8442 150.5258 144.4619 138.6424 133.0572
#>   [6,] 181.2044 173.9047 166.8990 160.1756 153.7230 147.5303 141.5871 135.8834
#>   [7,] 185.0532 177.5985 170.4440 163.5777 156.9881 150.6639 144.5945 138.7696
#>   [8,] 188.9838 181.3707 174.0642 167.0521 160.3225 153.8640 147.6657 141.7170
#>   [9,] 188.9838 185.2230 177.7614 170.6004 163.7278 157.1321 150.8021 144.7271
#>  [10,] 189.5921 185.2230 181.5371 174.2239 167.2054 160.4696 154.0052 147.8012
#>  [11,] 190.0921 185.8192 181.5371 177.9245 170.7569 163.8780 157.2763 150.9405
#>  [12,] 190.4844 186.3092 182.1214 177.9245 174.3838 167.3588 160.6169 154.1465
#>  [13,] 190.7704 186.6938 182.6017 178.4972 174.3838 170.9136 164.0284 157.4206
#>  [14,] 190.9518 186.9740 182.9786 178.9679 174.9451 170.9136 167.5124 160.7642
#>  [15,] 191.0310 187.1518 183.2533 179.3373 175.4064 171.4637 167.5124 164.1789
#>  [16,] 191.0112 187.2295 183.4275 179.6065 175.7685 171.9158 168.0515 164.1789
#>  [17,] 190.8958 187.2101 183.5037 179.7773 176.0323 172.2707 168.4947 164.7073
#>  [18,] 190.6890 187.0970 183.4846 179.8519 176.1997 172.5293 168.8425 165.1417
#>  [19,] 190.3954 186.8943 183.3738 179.8333 176.2729 172.6934 169.0960 165.4825
#>  [20,] 190.0201 186.6066 183.1751 179.7246 176.2546 172.7651 169.2568 165.7310
#>  [21,] 189.5685 186.2387 182.8931 179.5299 176.1481 172.7471 169.3270 165.8885
#>  [22,] 189.0465 185.7961 182.5325 179.2535 175.9573 172.6428 169.3094 165.9574
#>  [23,] 188.4602 185.2845 182.0987 178.9001 175.6864 172.4557 169.2072 165.9402
#>  [24,] 187.8161 184.7099 181.5973 178.4750 175.3400 172.1902 169.0239 165.8400
#>  [25,] 187.1206 184.0786 181.0342 177.9835 174.9233 171.8508 168.7636 165.6603
#>  [26,] 186.3805 183.3969 180.4154 177.4316 174.4417 171.4424 168.4309 165.4052
#>  [27,] 185.4469 182.6715 179.7473 176.8251 173.9007 170.9703 168.0307 165.0792
#>  [28,] 184.3727 181.7565 179.0364 176.1704 173.3063 170.4401 167.5680 164.6868
#>  [29,] 183.1787 180.7037 178.1395 175.4735 172.6646 169.8575 167.0483 164.2334
#>  [30,] 181.8856 179.5335 177.1076 174.5946 171.9816 169.2286 166.4774 163.7241
#>  [31,] 180.5132 178.2661 175.9608 173.5832 171.1201 168.5592 165.8609 163.1645
#>  [32,] 179.0804 176.9210 174.7186 172.4591 170.1289 167.7148 165.2049 162.5603
#>  [33,] 177.6052 175.5167 173.4002 171.2417 169.0272 166.7433 164.3773 161.9173
#>  [34,] 176.1041 174.0708 172.0239 169.9496 167.8340 165.6636 163.4251 161.1062
#>  [35,] 174.5923 172.5996 170.6068 168.6006 166.5676 164.4941 162.3669 160.1730
#>  [36,] 173.0835 171.1179 169.1649 167.2118 165.2455 163.2529 161.2207 159.1358
#>  [37,] 171.5897 169.6391 167.7127 165.7985 163.8842 161.9571 160.0041 158.0124
#>  [38,] 170.1214 168.1751 166.2633 164.3752 162.4991 160.6229 158.7342 156.8201
#>  [39,] 168.6876 166.7360 164.8284 162.9547 161.1041 159.2654 157.4265 155.5753
#>  [40,] 167.2957 165.3307 163.4179 161.5483 159.7119 157.8982 156.0960 154.2938
#>  [41,] 165.9518 163.9665 162.0406 160.1659 158.3335 156.5336 154.7560 152.9897
#>  [42,] 164.6604 162.6493 160.7036 158.8160 156.9786 155.1826 153.4186 151.6764
#>  [43,] 163.4252 161.3837 159.4126 157.5056 155.6556 153.8548 152.0945 150.3656
#>  [44,] 162.2484 160.1730 158.1721 156.2403 154.3712 152.5580 150.7930 149.0678
#>  [45,] 161.1275 159.0197 156.9856 155.0245 153.1311 151.2992 149.5221 147.7923
#>  [46,] 160.0610 157.9211 155.8552 153.8616 151.9395 150.0838 148.2884 146.5467
#>  [47,] 159.0474 156.8758 154.7785 152.7537 150.7997 148.9159 147.0971 145.3374
#>  [48,] 158.0854 155.8824 153.7539 151.6984 149.7139 147.7988 145.9525 144.1699
#>  [49,] 157.1736 154.9395 152.7803 150.6942 148.6796 146.7346 144.8576 143.0481
#>  [50,] 156.3106 154.0458 151.8562 149.7400 147.6954 145.7208 143.8146 141.9749
#>  [51,] 155.4951 153.2000 150.9803 148.8342 146.7601 144.7563 142.8210 140.9526
#>  [52,] 154.7256 152.4008 150.1514 147.9758 145.8724 143.8396 141.8756 139.9789
#>  [53,] 154.0007 151.6466 149.3680 147.1633 145.0311 142.9696 140.9772 139.0523
#>  [54,] 153.3188 150.9361 148.6288 146.3956 144.2348 142.1450 140.1245 138.1718
#>  [55,] 152.6784 150.2678 147.9325 145.6711 143.4823 141.3645 139.3163 137.3360
#>  [56,] 152.0779 149.6401 147.2775 144.9886 142.7723 140.6270 138.5514 136.5439
#>  [57,] 151.5156 149.0515 146.6623 144.3466 142.1034 139.9311 137.8285 135.7942
#>  [58,] 150.9899 148.5004 146.0854 143.7437 141.4741 139.2755 137.1465 135.0857
#>  [59,] 150.4991 147.9852 145.5453 143.1783 140.8832 138.6588 136.5039 134.4173
#>  [60,] 150.0416 147.5042 145.0403 142.6489 140.3291 138.0796 135.8995 133.7875
#>  [61,] 149.6156 147.0557 144.5688 142.1540 139.8102 137.5365 135.3318 133.1951
#>  [62,] 149.2196 146.6383 144.1293 141.6919 139.3251 137.0280 134.7995 132.6387
#>  [63,] 148.8517 146.2501 143.7202 141.2612 138.8723 136.5526 134.3011 132.1170
#>  [64,] 148.5104 145.8896 143.3397 140.8601 138.4501 136.1087 133.8352 131.6285
#>  [65,] 148.1938 145.5550 142.9864 140.4873 138.0570 135.6949 133.4001 131.1718
#>  [66,] 147.9004 145.2447 142.6585 140.1410 137.6916 135.3097 132.9946 130.7455
#>  [67,] 147.6285 144.9571 142.3544 139.8196 137.3522 134.9515 132.6170 130.3480
#>  [68,] 147.3767 144.6907 142.0725 139.5215 137.0372 134.6188 132.2660 129.9779
#>  [69,] 147.1436 144.4439 141.8113 139.2453 136.7450 134.3101 131.9399 129.6339
#>  [70,] 146.9278 144.2154 141.5695 138.9893 136.4743 134.0238 131.6374 129.3143
#>  [71,] 146.7281 144.0039 141.3455 138.7522 136.2234 133.7584 131.3567 129.0178
#>  [72,] 146.5433 143.8082 141.1382 138.5328 135.9911 133.5126 131.0966 128.7427
#>  [73,] 146.3723 143.6271 140.9464 138.3296 135.7760 133.2848 130.8557 128.4878
#>  [74,] 146.2142 143.4595 140.7689 138.1416 135.5768 133.0740 130.6325 128.2516
#>  [75,] 146.0680 143.3046 140.6047 137.9676 135.3925 132.8788 130.4258 128.0329
#>  [76,] 145.9329 143.1613 140.4528 137.8067 135.2220 132.6982 130.2346 127.8304
#>  [77,] 145.8079 143.0288 140.3124 137.6578 135.0643 132.5311 130.0575 127.6429
#>  [78,] 145.6924 142.9063 140.1825 137.5201 134.9184 132.3765 129.8937 127.4694
#>  [79,] 145.5858 142.7932 140.0625 137.3929 134.7835 132.2335 129.7422 127.3089
#>  [80,] 145.4873 142.6886 139.9516 137.2752 134.6588 132.1013 129.6021 127.1604
#>  [81,] 145.3963 142.5921 139.8491 137.1665 134.5435 131.9791 129.4725 127.0230
#>  [82,] 145.3123 142.5029 139.7545 137.0661 134.4369 131.8661 129.3527 126.8960
#>  [83,] 145.2348 142.4206 139.6671 136.9734 134.3385 131.7616 129.2419 126.7786
#>  [84,] 145.1633 142.3446 139.5864 136.8877 134.2476 131.6652 129.1396 126.6700
#>  [85,] 145.0974 142.2746 139.5120 136.8087 134.1636 131.5761 129.0450 126.5697
#>  [86,] 145.0365 142.2099 139.4433 136.7357 134.0862 131.4938 128.9577 126.4770
#>  [87,] 144.9804 142.1503 139.3799 136.6684 134.0147 131.4178 128.8771 126.3914
#>  [88,] 144.9287 142.0953 139.3215 136.6063 133.9487 131.3478 128.8026 126.3124
#>  [89,] 144.8810 142.0446 139.2676 136.5490 133.8878 131.2831 128.7339 126.2395
#>  [90,] 144.8370 141.9979 139.2179 136.4962 133.8317 131.2234 128.6706 126.1721
#>  [91,] 144.7965 141.9548 139.1721 136.4475 133.7799 131.1684 128.6121 126.1100
#>  [92,] 144.7591 141.9150 139.1299 136.4026 133.7322 131.1177 128.5582 126.0527
#>  [93,] 144.7246 141.8784 139.0909 136.3612 133.6882 131.0709 128.5084 125.9999
#>  [94,] 144.6929 141.8446 139.0550 136.3230 133.6476 131.0278 128.4626 125.9511
#>  [95,] 144.6636 141.8135 139.0219 136.2878 133.6102 130.9880 128.4203 125.9062
#>  [96,] 144.6366 141.7848 138.9914 136.2554 133.5757 130.9513 128.3813 125.8647
#>  [97,] 144.6117 141.7583 138.9633 136.2255 133.5439 130.9175 128.3454 125.8265
#>  [98,] 144.5888 141.7340 138.9374 136.1979 133.5146 130.8864 128.3123 125.7913
#>  [99,] 144.5677 141.7115 138.9135 136.1725 133.4876 130.8576 128.2817 125.7589
#> [100,] 144.5482 141.6908 138.8914 136.1491 133.4627 130.8312 128.2536 125.7289
#>           [,17]    [,18]    [,19]    [,20]     [,21]     [,22]     [,23]
#>   [1,] 117.4004 112.6709 108.1320 103.7760  99.59543  95.58328  91.73275
#>   [2,] 119.8940 115.0641 110.4288 105.9802 101.71086  97.61348  93.68117
#>   [3,] 122.4405 117.5081 112.7743 108.2313 103.87121  99.68681  95.67098
#>   [4,] 125.0412 120.0040 115.1697 110.5301 106.07745 101.80417  97.70304
#>   [5,] 127.6971 122.5529 117.6159 112.8778 108.33055 103.96651  99.77827
#>   [6,] 130.4094 125.1559 120.1141 115.2753 110.63151 106.17478 101.89758
#>   [7,] 133.1793 127.8142 122.6653 117.7238 112.98134 108.42994 104.06190
#>   [8,] 136.0080 130.5290 125.2707 120.2243 115.38108 110.73301 106.27219
#>   [9,] 138.8969 133.3015 127.9315 122.7778 117.83180 113.08500 108.52943
#>  [10,] 141.8471 136.1328 130.6488 125.3857 120.33456 115.48694 110.83461
#>  [11,] 144.8599 139.0243 133.4238 128.0489 122.89049 117.93990 113.18875
#>  [12,] 147.9368 141.9772 136.2577 130.7687 125.50070 120.44497 115.59290
#>  [13,] 151.0790 144.9928 139.1519 133.5462 128.16636 123.00324 118.04811
#>  [14,] 154.2879 148.0725 142.1075 136.3827 130.88863 125.61584 120.55547
#>  [15,] 157.5650 151.2176 145.1259 139.2795 133.66872 128.28395 123.11609
#>  [16,] 160.9117 154.4295 148.2084 142.2379 136.50787 131.00872 125.73109
#>  [17,] 160.9117 157.7096 151.3563 145.2590 139.40732 133.79136 128.40164
#>  [18,] 161.4296 157.7096 154.5712 148.3443 142.36835 136.63311 131.12891
#>  [19,] 161.8554 158.2172 154.5712 151.4952 145.39228 139.53522 133.91411
#>  [20,] 162.1894 158.6344 155.0687 151.4952 148.48043 142.49897 136.75847
#>  [21,] 162.4329 158.9619 155.4776 151.9828 148.48043 145.52567 139.66324
#>  [22,] 162.5874 159.2005 155.7985 152.3836 148.95834 145.52567 142.62971
#>  [23,] 162.6549 159.3519 156.0324 152.6981 149.35117 145.99407 142.62971
#>  [24,] 162.6380 159.4180 156.1808 152.9274 149.65944 146.37908 143.08879
#>  [25,] 162.5397 159.4015 156.2456 153.0728 149.88411 146.68121 143.46614
#>  [26,] 162.3637 159.3052 156.2294 153.1363 150.02663 146.90141 143.76226
#>  [27,] 162.1137 159.1326 156.1350 153.1204 150.08891 147.04110 143.97808
#>  [28,] 161.7941 158.8876 155.9659 153.0279 150.07333 147.10214 144.11498
#>  [29,] 161.4096 158.5744 155.7257 152.8622 149.98268 147.08687 144.17481
#>  [30,] 160.9651 158.1975 155.4188 152.6268 149.82021 146.99803 144.15984
#>  [31,] 160.4659 157.7619 155.0494 152.3259 149.58953 146.83878 144.07277
#>  [32,] 159.9175 157.2727 154.6224 151.9639 149.29463 146.61270 143.91669
#>  [33,] 159.3253 156.7351 154.1429 151.5455 148.93983 146.32367 143.69510
#>  [34,] 158.6951 156.1548 153.6161 151.0755 148.52971 145.97593 143.41183
#>  [35,] 157.9002 155.5371 153.0473 150.5591 148.06910 145.57397 143.07100
#>  [36,] 156.9855 154.7580 152.4419 150.0016 147.56301 145.12252 142.67704
#>  [37,] 155.9690 153.8615 151.6783 149.4083 147.01660 144.62651 142.23459
#>  [38,] 154.8679 152.8652 150.7997 148.6599 146.43509 144.09097 141.74844
#>  [39,] 153.6993 151.7860 149.8232 147.7988 145.70156 143.52103 141.22356
#>  [40,] 152.4794 150.6407 148.7655 146.8417 144.85757 142.80209 140.66496
#>  [41,] 151.2233 149.4451 147.6430 145.8051 143.91953 141.97491 139.96033
#>  [42,] 149.9452 148.2140 146.4711 144.7049 142.90355 141.05553 139.14961
#>  [43,] 148.6580 146.9613 145.2645 143.5563 141.82525 140.05977 138.24853
#>  [44,] 147.3733 145.6997 144.0368 142.3737 140.69956 139.00293 137.27258
#>  [45,] 146.1014 144.4406 142.8003 141.1704 139.54051 137.89964 136.23677
#>  [46,] 144.8512 143.1940 141.5662 139.9585 138.36114 136.76365 135.15543
#>  [47,] 143.6304 141.9687 140.3444 138.7490 137.17337 135.60776 134.04206
#>  [48,] 142.4452 140.7721 139.1435 137.5515 135.98792 134.44362 132.90916
#>  [49,] 141.3009 139.6106 137.9708 136.3745 134.81427 133.28176 131.76819
#>  [50,] 140.2014 138.4890 136.8323 135.2251 133.66067 132.13146 130.62945
#>  [51,] 139.1496 137.4114 135.7331 134.1094 132.53416 131.00083 129.50205
#>  [52,] 138.1477 136.3806 134.6769 133.0320 131.44058 129.89673 128.39391
#>  [53,] 137.1933 135.3985 133.6666 131.9968 130.38467 128.82492 127.31179
#>  [54,] 136.2852 134.4631 132.7041 131.0066 129.37009 127.79001 126.26130
#>  [55,] 135.4221 133.5731 131.7873 130.0633 128.39960 126.79563 125.24699
#>  [56,] 134.6030 132.7272 130.9150 129.1648 127.47505 125.84445 124.27240
#>  [57,] 133.8267 131.9244 130.0860 128.3098 126.59438 124.93829 123.34014
#>  [58,] 133.0919 131.1635 129.2991 127.4973 125.75642 124.07515 122.45202
#>  [59,] 132.3975 130.4434 128.5534 126.7261 124.96006 123.25387 121.60605
#>  [60,] 131.7424 129.7628 127.8475 125.9951 124.20422 122.47336 120.80112
#>  [61,] 131.1251 129.1207 127.1805 125.3034 123.48784 121.73256 120.03614
#>  [62,] 130.5445 128.5157 126.5512 124.6496 122.80984 121.03043 119.31008
#>  [63,] 129.9992 127.9467 125.9583 124.0328 122.16912 120.36592 118.62193
#>  [64,] 129.4879 127.4122 125.4005 123.4517 121.56456 119.73795 117.97064
#>  [65,] 129.0091 126.9111 124.8767 122.9051 120.99501 119.14542 117.35517
#>  [66,] 128.5615 126.4418 124.3856 122.3917 120.45926 118.58720 116.77443
#>  [67,] 128.1436 126.0031 123.9257 121.9103 119.95610 118.06212 116.22732
#>  [68,] 127.7541 125.5936 123.4957 121.4595 119.48427 117.56898 115.71269
#>  [69,] 127.3914 125.2118 123.0943 121.0381 119.04249 117.10654 115.22935
#>  [70,] 127.0542 124.8563 122.7200 120.6447 118.62946 116.67354 114.77612
#>  [71,] 126.7410 124.5258 122.3717 120.2779 118.24386 116.26873 114.35174
#>  [72,] 126.4503 124.2188 122.0477 119.9365 117.88439 115.89081 113.95498
#>  [73,] 126.1808 123.9340 121.7469 119.6190 117.54972 115.53849 113.58458
#>  [74,] 125.9309 123.6698 121.4677 119.3241 117.23856 115.21048 113.23927
#>  [75,] 125.6994 123.4249 121.2087 119.0505 116.94956 114.90552 112.91779
#>  [76,] 125.4850 123.1980 120.9687 118.7967 116.68136 114.62226 112.61890
#>  [77,] 125.2865 122.9879 120.7464 118.5615 116.43263 114.35940 112.34128
#>  [78,] 125.1028 122.7933 120.5404 118.3435 116.20208 114.11562 112.08365
#>  [79,] 124.9328 122.6133 120.3497 118.1417 115.98848 113.88966 111.84472
#>  [80,] 124.7754 122.4466 120.1732 117.9548 115.79065 113.68031 111.62325
#>  [81,] 124.6299 122.2924 120.0099 117.7818 115.60749 113.48642 111.41807
#>  [82,] 124.4952 122.1497 119.8588 117.6217 115.43794 113.30690 111.22804
#>  [83,] 124.3708 122.0178 119.7190 117.4736 115.28103 113.14073 111.05209
#>  [84,] 124.2557 121.8958 119.5896 117.3365 115.13585 112.98694 110.88923
#>  [85,] 124.1493 121.7830 119.4701 117.2098 115.00155 112.84465 110.73850
#>  [86,] 124.0510 121.6787 119.3595 117.0926 114.87733 112.71302 110.59904
#>  [87,] 123.9601 121.5823 119.2573 116.9842 114.76245 112.59127 110.47003
#>  [88,] 123.8763 121.4933 119.1629 116.8841 114.65625 112.47868 110.35070
#>  [89,] 123.7988 121.4111 119.0756 116.7915 114.55808 112.37459 110.24036
#>  [90,] 123.7273 121.3352 118.9950 116.7060 114.46737 112.27838 110.13834
#>  [91,] 123.6613 121.2651 118.9206 116.6270 114.38356 112.18947 110.04404
#>  [92,] 123.6004 121.2004 118.8519 116.5541 114.30615 112.10733 109.95690
#>  [93,] 123.5443 121.1408 118.7886 116.4868 114.23468 112.03146 109.87639
#>  [94,] 123.4925 121.0857 118.7301 116.4247 114.16870 111.96141 109.80203
#>  [95,] 123.4447 121.0350 118.6761 116.3673 114.10782 111.89675 109.73338
#>  [96,] 123.4007 120.9881 118.6264 116.3145 114.05164 111.83707 109.67000
#>  [97,] 123.3600 120.9450 118.5805 116.2657 113.99982 111.78201 109.61151
#>  [98,] 123.3226 120.9052 118.5382 116.2207 113.95202 111.73122 109.55755
#>  [99,] 123.2881 120.8685 118.4992 116.1793 113.90794 111.68437 109.50777
#> [100,] 123.2563 120.8346 118.4632 116.1410 113.86730 111.64117 109.46185
#>            [,24]     [,25]    [,26]
#>   [1,]  88.03735  84.49080 2012.861
#>   [2,]  89.90727  86.28540 2055.615
#>   [3,]  91.81692  88.11812 2099.277
#>   [4,]  93.76712  89.98976 2143.866
#>   [5,]  95.75875  91.90116 2189.402
#>   [6,]  97.79268  93.85315 2235.905
#>   [7,]  99.86982  95.84661 2283.396
#>   [8,] 101.99107  97.88241 2331.895
#>   [9,] 104.15737  99.96144 2381.425
#>  [10,] 106.36969 102.08464 2432.007
#>  [11,] 108.62900 104.25293 2483.663
#>  [12,] 110.93630 106.46728 2536.417
#>  [13,] 113.29260 108.72866 2590.291
#>  [14,] 115.69895 111.03808 2645.309
#>  [15,] 118.15642 113.39655 2701.496
#>  [16,] 120.66608 115.80511 2758.876
#>  [17,] 123.22905 118.26482 2817.475
#>  [18,] 125.84645 120.77679 2877.318
#>  [19,] 128.51945 123.34211 2938.433
#>  [20,] 131.24922 125.96191 3000.846
#>  [21,] 134.03698 128.63736 3064.584
#>  [22,] 136.88394 131.36964 3129.676
#>  [23,] 139.79138 134.15995 3196.151
#>  [24,] 139.79138 137.00953 3264.038
#>  [25,] 140.24132 137.00953 3333.367
#>  [26,] 140.61116 137.45052 3401.316
#>  [27,] 140.90139 137.81300 3468.345
#>  [28,] 141.11291 138.09745 3534.395
#>  [29,] 141.24709 138.30477 3599.410
#>  [30,] 141.30573 138.43628 3663.334
#>  [31,] 141.29106 138.49375 3726.115
#>  [32,] 141.20572 138.47937 3787.703
#>  [33,] 141.05275 138.39572 3848.052
#>  [34,] 140.83557 138.24580 3907.117
#>  [35,] 140.55793 138.03294 3964.860
#>  [36,] 140.22389 137.76083 4021.246
#>  [37,] 139.83777 137.43344 4076.242
#>  [38,] 139.40412 137.05500 4129.824
#>  [39,] 138.92765 136.62998 4181.968
#>  [40,] 138.41321 136.16299 4232.658
#>  [41,] 137.86573 135.65879 4281.881
#>  [42,] 137.17512 135.12220 4329.631
#>  [43,] 136.38053 134.44534 4375.904
#>  [44,] 135.49738 133.66656 4420.594
#>  [45,] 134.54086 132.80098 4463.631
#>  [46,] 133.52566 131.86349 4504.963
#>  [47,] 132.46584 130.86850 4544.553
#>  [48,] 131.37462 129.82977 4582.381
#>  [49,] 130.26427 128.76026 4618.438
#>  [50,] 129.14601 127.67201 4652.729
#>  [51,] 128.02993 126.57600 4685.271
#>  [52,] 126.92496 125.48213 4716.091
#>  [53,] 125.83887 124.39915 4745.226
#>  [54,] 124.77828 123.33468 4772.719
#>  [55,] 123.74870 122.29520 4798.623
#>  [56,] 122.75457 121.28610 4822.991
#>  [57,] 121.79937 120.31176 4845.886
#>  [58,] 120.88567 119.37557 4867.371
#>  [59,] 120.01522 118.48005 4887.510
#>  [60,] 119.18609 117.62692 4906.371
#>  [61,] 118.39717 116.81429 4924.020
#>  [62,] 117.64742 116.04107 4940.522
#>  [63,] 116.93581 115.30623 4955.938
#>  [64,] 116.26135 114.60879 4970.326
#>  [65,] 115.62302 113.94775 4983.745
#>  [66,] 115.01980 113.32212 4996.248
#>  [67,] 114.45062 112.73090 5007.890
#>  [68,] 113.91440 112.17305 5018.720
#>  [69,] 113.41000 111.64750 5028.789
#>  [70,] 112.93629 111.15314 5038.142
#>  [71,] 112.49207 110.68886 5046.824
#>  [72,] 112.07614 110.25348 5054.878
#>  [73,] 111.68728 109.84583 5062.345
#>  [74,] 111.32425 109.46470 5069.265
#>  [75,] 110.98581 109.10890 5075.673
#>  [76,] 110.67073 108.77719 5081.604
#>  [77,] 110.37778 108.46838 5087.093
#>  [78,] 110.10569 108.18126 5092.170
#>  [79,] 109.85318 107.91459 5096.864
#>  [80,] 109.61901 107.66710 5101.203
#>  [81,] 109.40195 107.43759 5105.214
#>  [82,] 109.20085 107.22485 5108.920
#>  [83,] 109.01460 107.02775 5112.343
#>  [84,] 108.84216 106.84521 5115.506
#>  [85,] 108.68253 106.67620 5118.426
#>  [86,] 108.53481 106.51975 5121.123
#>  [87,] 108.39812 106.37496 5123.612
#>  [88,] 108.27167 106.24100 5125.911
#>  [89,] 108.15472 106.11707 5128.032
#>  [90,] 108.04657 106.00244 5129.989
#>  [91,] 107.94658 105.89645 5131.796
#>  [92,] 107.85416 105.79845 5133.462
#>  [93,] 107.76876 105.70786 5134.999
#>  [94,] 107.68985 105.62416 5136.417
#>  [95,] 107.61697 105.54682 5137.724
#>  [96,] 107.54968 105.47540 5138.930
#>  [97,] 107.48757 105.40945 5140.042
#>  [98,] 107.43025 105.34857 5141.067
#>  [99,] 107.37735 105.29238 5142.012
#> [100,] 107.32856 105.24055 5142.883
#>