Get bycatch mortality rate given depletion

This function solves for the bycatch mortality rate \(E\) that gives a pre-specified depletion level InitDepl.w. It is used within the projections() function to calculate the stable age distribution at which to start the projections.

get_f(
  f.start = NA,
  S0.w = NA,
  S1plus.w = NA,
  nages.w = NA,
  AgeMat.w = NA,
  InitDepl.w = NA,
  z.w = NA,
  lambdaMax.w = NA,
  N0.w = NA,
  P0.w = NA,
  Check = FALSE
)

Arguments

f.start
an initial guess for the bycatch mortality rate E. The default value is E = 0.5

S0.w
Calf/pup survival, a numeric value between 0 and 1. (Note: the 'w' suffix indicates that \(z\) is in the wrapper function, and is used inside the function by optim)

S1plus.w
survival rate for animals age 1 year and older, a numeric value between 0 and 1

nages.w
"maximum" age, treated as the plus group age. The plus group age can be set equal to the age at maturity +2 years without losing accuracy.

AgeMat.w
Age at maturity in years (assumed to be age at first parturition - 1)

InitDepl.w
The depletion level to solve for as a proportion of carrying capacity; a numeric value between 0 and 1. This is equivalent to 'starting depletion'.

z.w
The degree of compensation. The default value is z = 2.39.

lambdaMax.w
The maximum intrinsic growth rate

N0.w
unfished numbers per recruit in terms of mature individuals individuals

P0.w
Number of individuals per recruit in terms of individuals aged 1+

Check
logical; if TRUE, prints the value to be minimized (should be zero)

Value

The bycatch rate that would lead to a depletion level of InitDepl.w -- a single value.

Examples

# Set parameters
S0.w = 0.5; S1plus.w = 0.944; 
nages.w = 25; AgeMat.w = 18 
InitDepl.w = 0.9; z.w = 2.39; lambdaMax.w = 1.04
# Get number of individuals per recruit in terms of mature individuals (N0.w)
NPROut <- npr(S0 = S0.w, S1plus = S1plus.w, 
nages = nages.w, AgeMat = AgeMat.w, E = 0)
N0 <- NPROut$npr # mature numbers per recruit
# Get number of individuals per recruit in terms of individuals aged 1+ (P0.w)
P0 <- NPROut$P1r # 1+ nums per recruit

# Get bycatch mortality rate for the initial depletion defined above
get_f(f.start = 0.5, 
S0.w = S0.w, S1plus.w = S1plus.w, nages.w = nages.w, AgeMat.w = AgeMat.w,
InitDepl.w = InitDepl.w, z.w = z.w, lambdaMax.w = lambdaMax.w, 
N0.w = N0, P0.w = P0, Check = FALSE)

#> [1] 0.01168733