Get relative recruitment at bycatch mortality rate E

calculates recruitment at bycatch mortality rate E, relative to recruitment for a no-bycatch scenario

get_rf(E_in, S0, S1plus, nages, AgeMat, z, A, P0, N0)

Arguments

E_in
Bycatch mortality rate, a numeric value between 0 and 1

S0
calf/pup survival, a numeric value between 0 and 1

S1plus
adult survival, a numeric value between 0 and 1

nages
number of age classes-- including plus group age-- in years

AgeMat
age at maturity in years, must be less than nages

z
degree of compensation, in the app calculated from the value of MNPL defined by the user

A
Pella-Tomlinson resilience parameter (see Punt 1999; Annex R). A = (FecMax - Fec0) / Fec0 (num)

P0
unfished 1+ numbers per recruit, \(\tildeP(0)\)

N0
unfished mature numbers per recruit, \(\tildeN(0)\)

Value

recruitment given exploitation rate E - this value is multiplied by the initial abundance \(N_{init}\) to get initial nums at age ( a vector)

Examples

S0 = 0.944; S1plus = 0.99; nages = 25; AgeMat = 17; z = 2.39; lambdaMax = 1.04;
NPROut <- npr(S0 = S0, S1plus = S1plus, nages = nages, AgeMat = AgeMat, E = 0)
N0 <- NPROut$npr # mature numbers per recruit
P0 <- NPROut$P1r # 1+ nums per recruit
Fec0 <- 1.0 / N0
FecMax <- getfecmax(S1plus = S1plus, S0 = S0, 
AgeMat = AgeMat, lambdaMax = lambdaMax)
A <- (FecMax - Fec0) / Fec0
get_rf(E_in = 0.01, S0 = S0, S1plus = S1plus, 
nages = nages, AgeMat = AgeMat, z = z, A = A,P0 = P0,N0 = N0)

#> [1] 1.845825