Run simulations of a marine mammal population with bycatch mortality

Generates several projections, stochasticity is in the number of catches from year to year

projections(
  NOut,
  ConstantBycatch = list(Catch = NA, CV = NA),
  ConstantRateBycatch = list(Rate = NA, CV = NA),
  InitDepl,
  lh.params,
  nyears,
  obs_CV = 0
)

Arguments

NOut
number of simulations

ConstantBycatch
Mean and CV of number of animals killed as bycatch per year (assumed lognormal)

ConstantRateBycatch
Mean and CV of bycatch rate (assumed normal)

InitDepl
initial depletion. If obs_CV>0, this is the mean depletion.

lh.params
- life history parameters as a list. The list must include S0, S1plus, K1plus, AgeMat, nages, z, and lambdaMax

nyears
number of years to do projections

obs_CV
observation CV. Default to 1 for simple projections

Value

list of outputs from simulations: params contains parameter values for each trajectory as a matrix; trajectories contains simulation outputs as a matrix; fishing.rates contain the bycatch rates for each year in each simulation as a matrix; InitDepl returns the initial depletion for the projections; ConstantBycatch provides Catch (total individuals killed in bycatch events per year) and CV of Catch (if the user has specified bycatch as a constant number); ConstantRateBycatch contains Bycatch Rate (additional mortality from bycatch each year) and CV of ByCatch rate. Other parameters are the same as in the dynamics() function.

Examples

projections(NOut = 3, ConstantRateBycatch = list(Rate = 0.01, CV = 0.3),
InitDepl = 0.8,
lh.params = list(S0=0.944,S1plus=0.99,
K1plus=9000,AgeMat=18,nages=20,z= 2.39,lambdaMax=1.02),
nyears = 50, obs_CV = 0)

#> $params
#>              S0 S1plus K1plus AgeMat nages    z lambdaMax
#> param.vec 0.944   0.99   9000     18    20 2.39      1.02
#> param.vec 0.944   0.99   9000     18    20 2.39      1.02
#> param.vec 0.944   0.99   9000     18    20 2.39      1.02
#> 
#> $trajectories
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> traj 7200 7180.390 7141.705 7113.274 7072.962 7072.143 7088.332 7049.556
#> traj 7200 7198.724 7219.979 7233.063 7230.308 7234.334 7217.176 7227.607
#> traj 7200 7237.060 7227.014 7263.854 7265.882 7290.727 7263.127 7234.473
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> traj 7045.522 7037.237 7051.839 7024.831 7061.659 7079.245 7100.990 7108.833
#> traj 7199.511 7191.731 7163.774 7173.559 7168.084 7164.485 7176.223 7169.137
#> traj 7251.943 7280.837 7285.209 7292.291 7297.611 7284.977 7285.350 7296.176
#>         [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#> traj 7123.704 7099.188 7081.389 7098.587 7070.215 7095.815 7076.627 7058.417
#> traj 7173.696 7169.289 7163.448 7135.714 7154.517 7167.596 7196.506 7199.747
#> traj 7236.240 7227.304 7224.768 7166.189 7175.926 7192.722 7199.289 7233.784
#>         [,25]    [,26]    [,27]    [,28]    [,29]    [,30]    [,31]    [,32]
#> traj 7045.663 7095.206 7102.514 7125.755 7130.195 7133.684 7105.390 7098.005
#> traj 7209.537 7196.743 7196.353 7161.761 7159.229 7164.339 7183.272 7173.031
#> traj 7235.645 7235.680 7216.153 7166.515 7173.296 7140.454 7165.244 7175.820
#>         [,33]    [,34]    [,35]    [,36]    [,37]    [,38]    [,39]    [,40]
#> traj 7108.611 7133.512 7128.227 7087.210 7056.683 7078.957 7108.275 7129.179
#> traj 7182.167 7134.255 7147.707 7160.125 7178.124 7151.159 7141.473 7137.860
#> traj 7198.311 7220.352 7160.962 7156.729 7144.715 7134.829 7150.729 7167.537
#>         [,41]    [,42]    [,43]    [,44]    [,45]    [,46]    [,47]    [,48]
#> traj 7124.855 7134.590 7141.479 7165.613 7158.205 7143.501 7122.434 7126.273
#> traj 7172.892 7212.160 7193.256 7193.539 7176.252 7190.046 7215.843 7203.860
#> traj 7168.171 7159.653 7130.423 7136.966 7096.477 7026.219 7052.818 7074.504
#>         [,49]    [,50]
#> traj 7164.437 7177.884
#> traj 7228.509 7205.033
#> traj 7063.388 7054.278
#> 
#> $fishing.rates
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> f.vec 0.012259613 0.015051892 0.013827729 0.015670020 0.010257603 0.007819659
#> f.vec 0.009687605 0.006532679 0.007574272 0.009723332 0.008793971 0.011734041
#> f.vec 0.004309327 0.010718153 0.004222933 0.008900251 0.005723928 0.012877256
#>              [,7]        [,8]        [,9]       [,10]       [,11]       [,12]
#> f.vec 0.015563253 0.010799709 0.011414970 0.008157912 0.014034194 0.004991166
#> f.vec 0.007971205 0.013304968 0.010618389 0.013493346 0.008330115 0.010426402
#> f.vec 0.013192636 0.006924456 0.005250503 0.008521176 0.008130047 0.008345212
#>             [,13]       [,14]       [,15]       [,16]      [,17]       [,18]
#> f.vec 0.007571824 0.006884727 0.008752822 0.007708536 0.01321659 0.012391791
#> f.vec 0.010187889 0.008040789 0.010629018 0.009023374 0.01025952 0.010482089
#> f.vec 0.010811016 0.009086105 0.007642053 0.017390491 0.01066435 0.009820021
#>             [,19]       [,20]       [,21]       [,22]       [,23]       [,24]
#> f.vec 0.007488383 0.013884684 0.006329717 0.012588343 0.012556306 0.011877047
#> f.vec 0.013596868 0.007163057 0.007874602 0.005572735 0.009037221 0.008098987
#> f.vec 0.017669668 0.008410341 0.007359508 0.008705322 0.004744682 0.009135477
#>             [,25]       [,26]       [,27]       [,28]       [,29]       [,30]
#> f.vec 0.003013159 0.008828762 0.006534319 0.009101335 0.009223525 0.013717053
#> f.vec 0.011212749 0.009540624 0.014342725 0.010023724 0.008957992 0.006983522
#> f.vec 0.009375552 0.012104132 0.016427309 0.008774762 0.014308081 0.006333480
#>             [,31]       [,32]       [,33]       [,34]      [,35]       [,36]
#> f.vec 0.010913033 0.008394734 0.006312541 0.010473161 0.01556496 0.014311770
#> f.vec 0.010996027 0.008324899 0.016303727 0.007907925 0.00798580 0.007136214
#> f.vec 0.008206751 0.006465158 0.006412458 0.017691757 0.01028435 0.011399812
#>             [,37]       [,38]       [,39]      [,40]       [,41]       [,42]
#> f.vec 0.006924496 0.005813539 0.006877459 0.01035610 0.008388678 0.008748922
#> f.vec 0.013377493 0.011092965 0.010286318 0.00483459 0.004078211 0.012050636
#> f.vec 0.011158536 0.007558009 0.007352376 0.00955021 0.010836125 0.013800306
#>             [,43]      [,44]       [,45]       [,46]       [,47]       [,48]
#> f.vec 0.006282920 0.01062316 0.011697112 0.012679055 0.009263260 0.004376618
#> f.vec 0.009461941 0.01192671 0.007646337 0.005892902 0.011063969 0.005997909
#> f.vec 0.008893156 0.01550837 0.019976880 0.006497485 0.007070198 0.011646396
#>             [,49]       [,50]
#> f.vec 0.007695963 0.007810230
#> f.vec 0.012609793 0.006380549
#> f.vec 0.011411859 0.008668286
#> 
#> $InitDepl
#> [1] 0.8
#> 
#> $ConstantBycatch
#> Catch    CV 
#>    NA    NA 
#> 
#> $ConstantRateBycatch
#> Rate   CV 
#> 0.01 0.30 
#>